Codeup2

+++ algorithm/Codeup1029.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 public class Codeup1029 {
     public static void main(String args[]){
         Scanner sc = new Scanner(System.in);
-        long data =sc.nextLong();
+        double data =sc.nextDouble();
 
         System.out.printf("%.11f",data);
         sc.close();
